refactor: Split ch3 into ch3 and ch4 (#65)
