+++ examples/ch06/counter/counter.html
