--- listings/ch03/listing_07.js
@@ -1,30 +1,0 @@
-import { DOM_TYPES } from './h'
-
-export function mountDOM(vdom, parentEl) {
-  switch (vdom.type) {
-    case DOM_TYPES.TEXT: {
-      createTextNode(vdom, parentEl) // --1--
-      break
-    }
-
-    case DOM_TYPES.ELEMENT: {
-      createElementNode(vdom, parentEl) // --2--
-      break
-    }
-
-    case DOM_TYPES.FRAGMENT: {
-      createFragmentNode(vdom, parentEl) // --3--
-      break
-    }
-
-    default: {
-      throw new Error(`Can't mount DOM of type: ${vdom.type}`)
-    }
-  }
-}
-
-// TODO: implement createTextNode()
-
-// TODO: implement createElementNode()
-
-// TODO: implement createFragmentNode()
