--- listings/ch03/listing_08.js
@@ -1,8 +1,0 @@
-function createTextNode(vdom, parentEl) {
-  const { value } = vdom
-
-  const textNode = document.createTextNode(value) //--1--
-  vdom.el = textNode //--2--
-
-  parentEl.append(textNode) //--3--
-}
