--- listings/ch03/listing_09.js
@@ -1,9 +1,0 @@
-function createFragmentNode(vdom, parentEl) {
-  const { children } = vdom
-
-  const fragment = document.createDocumentFragment() //--1--
-  vdom.el = parentEl //--2--
-
-  children.forEach((child) => mountDOM(child, fragment)) //--3--
-  parentEl.append(fragment) //--4--
-}
