--- listings/ch03/listing_11.js
@@ -1,4 +1,0 @@
-export function addEventListener(eventName, handler, el) {
-  el.addEventListener(eventName, handler)
-  return handler
-}
