--- listings/ch03/listing_12.js
@@ -1,10 +1,0 @@
-export function addEventListeners(listeners = {}, el) {
-  const addedListeners = {}
-
-  Object.entries(listeners).forEach(([eventName, handler]) => {
-    const listener = addEventListener(eventName, handler, el)
-    addedListeners[eventName] = listener
-  })
-
-  return addedListeners
-}
