--- listings/ch03/listing_15.js
@@ -1,7 +1,0 @@
-export function setStyle(el, name, value) {
-  el.style[name] = value
-}
-
-export function removeStyle(el, name) {
-  el.style[name] = null
-}
