--- listings/ch03/listing_18.js
@@ -1,4 +1,0 @@
-function removeTextNode(vdom) {
-  const { el } = vdom
-  el.remove()
-}
