--- listings/ch03/listing_19.js
@@ -1,11 +1,0 @@
-function removeElementNode(vdom) {
-  const { el, children, listeners } = vdom
-
-  el.remove()
-  children.forEach(destroyDOM)
-
-  if (listeners) {
-    removeEventListeners(listeners, el)
-    delete vdom.listeners
-  }
-}
