--- listings/ch03/listing_20.js
@@ -1,5 +1,0 @@
-export function removeEventListeners(listeners = {}, el) {
-  Object.entries(listeners).forEach(([eventName, handler]) => {
-    el.removeEventListener(eventName, handler)
-  })
-}
