--- listings/ch03/listing_21.js
@@ -1,4 +1,0 @@
-function removeFragmentNode(vdom) {
-  const { children } = vdom
-  children.forEach(destroyDOM)
-}
