+++ listings/ch04/listing_09.js
@@ -1,46 +1,7 @@
-import { destroyDOM } from './destroy-dom'
-import { Dispatcher } from './dispatcher'
-import { mountDOM } from './mount-dom'
+export function setStyle(el, name, value) {
+  el.style[name] = value
+}
 
-export function createApp({ state, view, reducers = {} }) {
-  let parentEl = null
-  let vdom = null
-
-  const dispatcher = new Dispatcher()
-  const subscriptions = [dispatcher.afterEveryCommand(renderApp)]
-
-  function emit(eventName, payload) {
-    dispatcher.dispatch(eventName, payload)
-  }
-
-  for (const actionName in reducers) {
-    const reducer = reducers[actionName]
-
-    const subs = dispatcher.subscribe(actionName, (payload) => {
-      state = reducer(state, payload)
-    })
-    subscriptions.push(subs)
-  }
-
-  function renderApp() {
-    if (vdom) {
-      destroyDOM(vdom)
-    }
-
-    vdom = view(state, emit)
-    mountDOM(vdom, parentEl)
-  }
-
-  return {
-    mount(_parentEl) {
-      parentEl = _parentEl
-      renderApp()
-    },
-
-    unmount() {
-      destroyDOM(vdom)
-      vdom = null
-      subscriptions.forEach((unsubscribe) => unsubscribe())
-    },
-  }
+export function removeStyle(el, name) {
+  el.style[name] = null
 }
