+++ listings/ch04/listing_14.js
@@ -1,21 +1,5 @@
-function CreateTodo({ currentTodo }, emit) {
-  return h('div', {}, [
-    h('label', { for: 'todo-input' }, ['New TODO']),
-    h('input', {
-      type: 'text',
-      id: 'todo-input',
-      value: currentTodo,
-      on: {
-        input: ({ target }) => emit('update-current-todo', target.value),
-      },
-    }),
-    h(
-      'button',
-      {
-        disabled: currentTodo.length < 3,
-        on: { click: () => emit('add-todo') },
-      },
-      ['Add']
-    ),
-  ])
+export function removeEventListeners(listeners = {}, el) {
+  Object.entries(listeners).forEach(([eventName, handler]) => {
+    el.removeEventListener(eventName, handler)
+  })
 }
