+++ listings/ch04/listing_15.js
@@ -1,7 +1,4 @@
-function TodoList({ todos, edit }, emit) {
-  return h(
-    'ul',
-    {},
-    todos.map((todo, i) => TodoItem({ todo, i, edit }, emit))
-  )
+function removeFragmentNode(vdom) {
+  const { children } = vdom
+  children.forEach(destroyDOM)
 }
