+++ listings/ch05/listing_04.js
@@ -1,0 +1,41 @@
+export class Dispatcher {
+  #subs = new Map()
+  #afterHandlers = []
+
+  subscribe(commandName, handler) {
+    if (!this.#subs.has(commandName)) {
+      this.#subs.set(commandName, [])
+    }
+
+    const handlers = this.#subs.get(commandName)
+    if (handlers.includes(handler)) {
+      return () => {}
+    }
+
+    handlers.push(handler)
+
+    return () => {
+      const idx = handlers.indexOf(handler)
+      handlers.splice(idx, 1)
+    }
+  }
+
+  afterEveryCommand(handler) {
+    this.#afterHandlers.push(handler)
+
+    return () => {
+      const idx = this.#afterHandlers.indexOf(handler)
+      this.#afterHandlers.splice(idx, 1)
+    }
+  }
+
+  dispatch(commandName, payload) {
+    if (this.#subs.has(commandName)) {
+      this.#subs.get(commandName).forEach((handler) => handler(payload))
+    } else {
+      console.warn(`No handlers for command: ${commandName}`)
+    }
+
+    this.#afterHandlers.forEach((handler) => handler())
+  }
+}
