+++ listings/ch06/listing_02.js
@@ -1,0 +1,15 @@
+import {
+  createApp,
+  h,
+  hFragment,
+} from '../../packages/runtime/dist/_<fwk-name>_'
+
+const state = {
+  currentTodo: '',
+  edit: {
+    idx: null,
+    original: null,
+    edited: null,
+  },
+  todos: ['Walk the dog', 'Water the plants'],
+}
