+++ listings/ch06/listing_04.js
@@ -1,0 +1,7 @@
+function App(state, emit) {
+  return hFragment([
+    h('h1', {}, ['My TODOs']),
+    CreateTodo(state, emit),
+    TodoList(state, emit),
+  ])
+}
