+++ listings/ch06/listing_05.js
@@ -1,0 +1,21 @@
+function CreateTodo({ currentTodo }, emit) {
+  return h('div', {}, [
+    h('label', { for: 'todo-input' }, ['New TODO']),
+    h('input', {
+      type: 'text',
+      id: 'todo-input',
+      value: currentTodo,
+      on: {
+        input: ({ target }) => emit('update-current-todo', target.value),
+      },
+    }),
+    h(
+      'button',
+      {
+        disabled: currentTodo.length < 3,
+        on: { click: () => emit('add-todo') },
+      },
+      ['Add']
+    ),
+  ])
+}
