+++ listings/ch06/listing_06.js
@@ -1,0 +1,7 @@
+function TodoList({ todos, edit }, emit) {
+  return h(
+    'ul',
+    {},
+    todos.map((todo, i) => TodoItem({ todo, i, edit }, emit))
+  )
+}
