+++ listings/ch06/listing_07.js
